src/requirements.txt
django==3.2.16
django-environ==0.9.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-seed==0.3.1
pymysql==1.0.2
djangorestframework-simplejwt==5.2.2
pillow==9.4.0
bcrypt==4.0.1
mastercard-oauth1-signer==1.6.1
requests==2.28.2
dict_hash==1.1.29
numpy==1.21.6
mercadopago==2.2.0
gunicorn
django-apscheduler==0.6.2
filestack-python==3.5.0